Infrastructure: Fix code signing to only run on the main repo (#7581)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Fix code signing to only run on the main repo - so far it either tried
to run on forks or refused to run in the main repo.
#### Motivation for adding to Mudlet
Sign builds for Mudlet only when the Azure secret is available.
#### Other info (issues closed, discussion etc)

(cherry picked from commit cdfd8341445f80760495a2451e171c49c6edf178)